fix: more utils move to util.ts
