+++ src/util.ts
@@ -8,6 +8,10 @@ export type Expand<T> = T extends object
     : never
   : T
 
+export type MaybeArray<T> = T | T[]
+export type MaybePromise<T> = T | Promise<T>
+export type NonemptyReadonlyArray<T> = readonly [T, ...T[]]
+
 type MaybeExtra<Extra> = (Extra & { caption?: string }) | undefined
 
 export function fmtCaption<
@@ -27,6 +31,14 @@ export function fmtCaption<
     }),
   }
 }
+
+/** Construct a generic type guard */
+export type Guard<X = unknown, Y extends X = X> = (x: X) => x is Y
+
+/** Extract the guarded type from a type guard, defaults to never. */
+export type Guarded<F> =
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  F extends (x: any) => x is infer T ? T : never
 
 export function deprecate(
   method: string,
